check base url

+++ install.sh
@@ -1,8 +1,19 @@
 #!/bin/bash
+
+# Check if the current directory is not equal to $HOME
+if [ "$PWD" != "$HOME" ]; then
+    echo "Install only in your home directory ($HOME). Stopping installer..."
+    exit 1
+fi
+
 tag="main"
 zipfile="$tag.zip"
 dir="kioskscripts-$tag/"
 kioskHome="$HOME/kiosk/"
 rm "$zipfile" && rm -r "$dir" && rm -rf "$kioskHome"
 mkdir "$kioskHome"
-wget "https://github.com/dsentker/kioskscripts/archive/$tag.zip" && unzip "$zipfile" && mv -v "$dir"* "$kioskHome" && cd "$kioskHome" && ./prepare.sh
+wget -q "https://github.com/dsentker/kioskscripts/archive/$tag.zip"
+unzip "$zipfile"
+mv -v "$dir"* "$kioskHome"
+cd "$kioskHome"
+./prepare.sh
